requirements.txt
flask
mysql-connector-python
python-dotenv
flask-sqlalchemy
pymysql
flask-marshmallow
passlib